TASK_#141

+++ contracts/interfaces/IWrappedNativeToken.sol
@@ -10,6 +10,6 @@ interface IWrappedNativeToken is IERC20 {
 
     function deposit() external payable;
 
-    function withdraw(uint wad) external;
+    function withdraw(uint256 wad) external;
 
 }
